Add host list command
We already have the "deployment list-hosts", but that requires the the
user specify the deployment ID. In normal installations, that
shouldn't be required because there is exactly one deployment.

In addition, it's more natural to be able to do "host list" since we
already have other commands like "host create" and "host show". So
this command make sense and should exist.

For now, I've left "deployment list-hosts" here for backwards
compatibility. We can consider removing it in the future if we like.

On my Mac, "make test" found a few minor error that were not found in the
Jenkins builds. I don't know why they weren't found, but they were
reasonable to fix, and they allowed me to run "make test", so I fixed
them as part of this checkin.

Change-Id: Iada2fc8ef21a963d75308ce373d589e410f51684