Don't print bad timestamps for incomplete tasks
Tasks print starting and ending timestamps. When a task is created,
the ending timestamp is "0". The CLI printed that as a date, which
made it look like the task ended at the beginning of the Unix epoch
(12:00am GMT January 1, 1970).

This change prints those timestamps as a hyphen ("-"), to avoid
confusion.

This only affects the interactive mode (not the scripting mode), so it
won't affect our tests.

While I was at it, I fixed a couple of typos in networks_test.go.

Change-Id: I3f66f8ecef0c1796f2bf03b4096bdccfb5db6a14