Add --output=json global option for images
This is the first step of several to add customized output options for
the photon CLI. This adds a global option, --output=json, but to keep
the change reasonable it is only implemented for the images
subcommand.

After this is committed, the plan is to add more types of output
(jsonpath for output single fields, and table for outputting custom
tables based on jsonpaths). Also, we'll expand coverage to all the
subcommands.

Specific changes:
- Added utils/formatting.go. All custom output happens here.
- Updated images.go to use the new formatting. Please notice that the
  "image create" command was updated to output the newly created
  image. This requires a GET of the image after it's created.
- Updated images_tests.go with basic coverage. This required minor
  refactoring in images.go to support checking the output.
- Updated waitOnTaskOperation() to be passed the command-line args
  instead of whether or not we're in scripting mode. This allows us to
  support the above case (image get) by not outputting anything when
  using custom output
- Updated printTaskList() to be passed the command-line args, so it
  can print the tasks with the custom formatting if appropriate.

The last two changes required touching a lot of files, but the changes
are trivial and improve the API for the callers.

Change-Id: Ia17546c4ea89bfd99ee2f39cdb04805058589b01