Add deployment command: set-security-groups
The CLI did not allow us to set security groups on a deployment, but
this is a capability we need so they can be modified after the
intial creation, similar to tenants and projects.

This was named set-security-groups instead of set_security_groups to
match the other commands with the deployment sub-command.

Change-Id: Ie31c830d45fb9c28fb1e4e117e981099f5638428