Added Option worker_count and batch_size_worker for CLI
This is an intermediate step for removing option slave.
Specifically, when I set worker_count, I picked the max of the worker_count and slave_count
1. For now, when i submit this code, Ruby test will test against slave_count, However, I changed
it in a way that even if user passed slave_count, i will pass worker_count to the go sdk. In the
CreateSpec I will create a spec with WorkerCount set.
2. After I change the Ruby test, Ruby will pass in worker count to cli, which still passes worker
Count to go-sdk. Either way, it should not break

Change-Id: I1ef1ce652ccec257c5599d11b587459c68c4a07a