Bugfix: gofmt error not caught in jenkin job
I have a gofmt error in my code when i commit to jenkin
When I run test locally, it was able to caught the gofmt
error and exit accordingly. However, when the test got
run in jenkin, it actually got passed without any error

The cause is that the jenkin machine we use remotely to run
the test have an OS that doesn't support "read" command
with out a parameter

I couldnt figure out an easy way to rewrite the makefile.
Thus, I added a script in ci folder called check-fmt.sh
which does the checking of gofmt and have the makefile
call the script instead.

Change-Id: I7de9a063015571bbf5c9ed1e1f655be102e6ea46