Fix: cluster show prints only json objects
Cluster show in json mode will output two json objects.
The first object is a cluster json object specified in
the go-sdk.
The second is a list of objects each containing the following
fields:
"vm" - a VM json object specified in the go-sdk
"ipAddress" - the corresponding IP address of the VM object.
The list of VMs will contain the ETCD and master VMs, but
none of the worker VMs.

Change-Id: I8411309bc934e29291bacfba4446e24c15b35b63