Delete GoCli Mesos and Swarm related command
For the photon controller 1.1 release, we are not
supporting Mesos and Swarm. Thus, I am removing all
supporting commands in the cli cluster commands for Mesos
and Swarm.

There is no need for a integration test change because
for the lifecycle test, we only use the api and not the cli

There is a go cli mesos and swarm ruby unit test in Ruby
folder ruby/common/spec/gocliclient/cluster_api_spec.rb
 in photon controller. I didn't delete that test, because
even though it was a unit test for go cli, everything was
a mock.

Change-Id: I0af3d1994ee0f6146294869fd1d025c524b3553c