Cli: Add a global option(--details) to replace "Using target ..."
Currently, if non-interactive mode is not set, whenever we do a
cli command, we see "Using target .." line. This line is not
very useful.

The change will:
a. by default, not print the using target info. For the output
   still print the header such as Name, ID etc for readability
b. If the user specifies a global option (--details or -d)
   we print a header message
Target: http://10.159.36.3:9000
Tenant: Tenant-A
Project: <not-set>
User: Token
c. This option should only be valid in the interactive mode:
   it's should be incompatible with -n or -o=json.
d. For non-interactive mode and -o=json mode, it wil not allow interaction.
   but for -detail mode and default mode, it will allow interaction

In unit test, it also no longer print the target info because unit tests are
either default or non-interactive mode

Change-Id: I9d6fb571540b3aa2dc4becd265ac3462aa676521