Bugfix: Validate Harbor admin-password input
Harbor require that the password to have at least 7 characters
with 1 lowercase letter, 1 capital letter and 1 numeric character.
However, the current harbor create command allow user to create
with any password. As a result, the harbor cluster would appear
to be ready and the user would not be able to log in to the cluster
using the admin-password created. Also, it would overwrite the
default Habor password so that the default password doesn't work
either.

This change basically add a sanity check to make sure that the
user admin account will be created correctly. Also some
documentation

Change-Id: I964dfd4687a404f8ed111e52549d8f985225b67e