Update codegangsta cli framework references to urfave cli
Replaces codegangsta cli in vendor and imports with urfave cli.
Govendor is updated to use urfave cli version 1.19.1
Also updates the ordering of import statements.

Urfave cli uses the same library as codegangsta,
but now the ownership has changed to urfave. This change will
improve clarity of references and enable easier updates to the
cli framework.

Change-Id: I6a5710226e91eba01a223899a7f5a05af46bf8c4