Add --boot-disk-flavor to vm create
When --boot-disk-flavor is present, the boot disk is inserted as the
first disk implicitly, and CLI may ask for additional data disks which
can no longer be a boot disk.

When --boot-disk-flavor is absent, in interactive mode,  it's prompted
and accepts empty value. If it's empty, disks are prompted as usually
and at least 1 boot disk is required.

Change-Id: I999fabaa41bc5a9502d367bdd1405a409acb9c6f