Add some commands to help user automate kubectl context
The output of photon service get-kubectl-auth will now also
include command to set the cluster, link the cluster and user
as a context and set the context as default.

As long as the user has kubectl in the path, the user can
now just copy the output of {photon service get-kubectl-auth
-u <username> -p <password> <service-id>} and it will configure
all the kubectl config. Alternatively, the user can now run
eval "$(photon service get-kubectl-auth -u <username> -p <password> <service-id>)"
to directly set the kubectl config

Change-Id: Ia374bb47c9f2b9575b779cd169576929e76af8ad